moved settings verification to main
